backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
xai-python==0.0.9
pinecone-client==3.0.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
spacy==3.7.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
feedparser==6.0.10
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4

